backend/requirements.txt
# SmartSense Backend Dependencies
# Install with: pip install -r requirements.txt

fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
websockets==12.0
requests==2.31.0
pyserial==3.5
